requirements.txt
boto3
requests
pyjwt
cryptography